pyproject.toml
[project]
name = "multi-agents-project"
version = "0.1.0"
description = ""
authors = [
    {name = "dkaddaf_expedia",email = "dkaddaf@expediagroup.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pandas (==2.2.3)",
    "typing (==3.7.4.3)",
    "sqlalchemy (==2.0.40)",
    "python-dotenv (==1.1.0)",
    "smolagents[litellm] (>=1.21.3,<2.0.0)",
    "openai (>=1.107.3,<2.0.0)",
    "pydantic-ai (>=1.0.7,<2.0.0)"
]

# Add this section to specify the package directory
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
